Minor Partition Refactor
