+++ src/main/scala/com/raphtory/core/actors/partitionmanager/workers/AnalysisSubtaskWorker.scala
@@ -9,9 +9,9 @@ import com.raphtory.core.actors.partitionmanager.workers.AnalysisSubtaskWorker.S
 import com.raphtory.core.actors.RaphtoryActor
 import com.raphtory.core.analysis.GraphLens
 import com.raphtory.core.analysis.api.Analyser
+import com.raphtory.core.model.GraphPartition
 import com.raphtory.core.model.communication.VertexMessage
 import com.raphtory.core.model.communication.VertexMessageHandler
-import com.raphtory.core.model.storage.GraphPartition
 import kamon.Kamon
 
 import scala.collection.mutable
