+++ src/main/scala/com/raphtory/core/analysis/entity/EntityVisitor.scala
@@ -1,7 +1,7 @@
 package com.raphtory.core.analysis.entity
 
 import com.raphtory.core.analysis.GraphLens
-import com.raphtory.core.model.implementations.entities.RaphtoryEntity
+import com.raphtory.core.model.implementations.objectgraph.entities.RaphtoryEntity
 
 import scala.collection.mutable
 import scala.collection.parallel.mutable.ParTrieMap
