+++ src/main/scala/com/raphtory/core/model/implementations/objectgraph/entities/RaphtoryVertex.scala
@@ -1,9 +1,9 @@
-package com.raphtory.core.model.implementations.entities
+package com.raphtory.core.model.implementations.objectgraph.entities
 
 import com.raphtory.core.actors.partitionmanager.workers.ParquetVertex
 import com.raphtory.core.analysis.GraphLens
 import com.raphtory.core.analysis.entity.{Edge, Vertex}
-import com.raphtory.core.model.storage.GraphPartition
+import com.raphtory.core.model.GraphPartition
 
 import scala.collection.mutable
 import scala.collection.parallel.mutable.ParTrieMap
