--- src/main/scala/com/raphtory/core/model/storage/AnalysisInterface.scala
@@ -1,5 +1,0 @@
-package com.raphtory.core.model.storage
-
-class AnalysisInterface {
-
-}
